feat: refactor google login to a component
